% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{filterWindows}
\alias{consolidateWindows}
\alias{consolidateTests}
\alias{consolidateOverlaps}
\title{Defunct functions}
\usage{
filterWindows(...)

consolidateWindows(...)

consolidateTests(...)

consolidateOverlaps(...)
}
\arguments{
\item{...}{Ignored arguments.}
}
\value{
All functions error out with a defunct message pointing towards its descendent (if available).
}
\description{
Functions that have passed on to the function afterlife.
Their successors are also listed.
}
\details{
\code{filterWindows} is succeeded by \code{\link{filterWindowsGlobal}} and related functions, which provide a more focused programmatic interface.

\code{consolidateWindows} is succeeded by \code{\link{mergeWindowsList}} and \code{\link{findOverlapsList}}.

\code{consolidateTests} and \code{consolidateOverlaps} are succeeded by \code{\link{mergeResultsList}} and \code{\link{overlapResultsList}}, respectively.
}
\examples{
try(filterWindows())
}
\author{
Aaron Lun
}
