% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalTests.R
\name{minimalTests}
\alias{minimalTests}
\title{Require rejection of a minimal number of tests}
\usage{
minimalTests(
  ids,
  tab,
  min.sig.n = 3,
  min.sig.prop = 0.4,
  weight = NULL,
  pval.col = NULL,
  fc.col = NULL
)
}
\arguments{
\item{ids}{An integer vector or factor containing the cluster ID for each test.}

\item{tab}{A data.frame of results with \code{PValue} and at least one \code{logFC} field for each test.}

\item{min.sig.n}{Integer scalar containing the minimum number of significant barcodes when \code{method="holm-min"}.}

\item{min.sig.prop}{Numeric scalar containing the minimum proportion of significant barcodes when \code{method="holm-min"}.}

\item{pval.col}{An integer scalar or string specifying the column of \code{tab} containing the p-values.}

\item{fc.col}{An integer or character vector specifying the columns of \code{tab} containing the log-fold changes.}
}
\value{
A \linkS4class{DataFrame} containing:
\itemize{
\item An integer field \code{NumTests}, specifying the total number of tests in each cluster.
\item One or more numeric fields containing the log-fold changes of the test with the \eqn{x}th-smallest p-value in each cluster.
\item A numeric field containing the combined p-value. 
If \code{pval.col=NULL}, this column is named \code{PValue}, otherwise its name is set to \code{colnames(tab[,pval.col])}.
\item A numeric field \code{FDR}, containing the q-value corresponding to the combined p-value.
}
}
\description{
Compute a p-value for each cluster based around the rejection of a minimal number or proportion of tests from that cluster.
}
\details{
For each cluster, this function applies the Holm-Bonferroni correction to all of its tests.
It then chooses the \eqn{x}th-smallest adjusted p-value as the cluster-level p-value,
where \eqn{x} is defined from the larger of \code{min.sig.n} and the product of \code{min.sig.prop} and the number of tests.
(If \eqn{x} is larger than the total number of tests, the largest per-test p-value is used instead.)
This means that a cluster can only achieve a low p-value if at least \eqn{x} tests also have low p-values.

More formally, the cluster-level null hypothesis is that the per-test null hypothesis is false for fewer than \eqn{x} tests.
We reject the cluster-level null at a nominal threshold \eqn{T} if we can reject \eqn{x} or more tests while controlling the FWER at \eqn{T}.
This is done using the Holm-Bonferroni method to obtain strong FWER control over the rejected tests in the cluster.
It is then straightforward to invert this logic to obtain a cluster-level p-value from the Holm-adjusted per-test p-values.
}
\examples{
ids <- round(runif(100, 1, 10))
tab <- data.frame(logFC=rnorm(100), logCPM=rnorm(100), PValue=rbeta(100, 1, 2))
minimal <- minimalTests(ids, tab)
head(minimal)

}
\author{
Aaron Lun
}
